+++ examples/service/verify/service/delete/service_delete_example.go
@@ -1,0 +1,36 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v2 "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var verifySession *v2.Verify
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	verifySession = twilio.NewWithCredentials(creds).Verify.V2
+}
+
+func main() {
+	err := verifySession.
+		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		Delete()
+
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	log.Println("Service deleted")
+}