+++ service/verify/tests/acceptance/verify_acceptance_suite_test.go
@@ -1,0 +1,35 @@
+// +build acceptance verify_acceptance
+
+package acceptance
+
+import (
+	"fmt"
+	"os"
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+// Prerequisites
+
+// 1) Twilio Account SID set as an environment variable - TWILIO_ACCOUNT_SID
+// 2) Twilio Auth Token set as an environment variable - TWILIO_AUTH_TOKEN
+
+func TestAcceptance(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Verify Acceptance Test Suite")
+}
+
+var _ = BeforeSuite(func() {
+	variables := []string{
+		"TWILIO_ACCOUNT_SID",
+		"TWILIO_AUTH_TOKEN",
+	}
+
+	for _, variable := range variables {
+		if value := os.Getenv(variable); value == "" {
+			Fail(fmt.Sprintf("`%s` are required for running acceptance tests", variable))
+		}
+	}
+})