+++ service/verify/tests/acceptance/verify_acceptance_test.go
@@ -1,0 +1,64 @@
+package acceptance
+
+import (
+	"fmt"
+	"os"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/services"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var _ = Describe("Verify Acceptance Tests", func() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		Fail(fmt.Sprintf("Failed to create credentials. Error %s", err.Error()))
+	}
+
+	verifySession := twilio.NewWithCredentials(creds).Verify.V2
+
+	Describe("Given the verify service clients", func() {
+		It("Then the service is created, fetched, updated and deleted", func() {
+			servicesClient := verifySession.Services
+
+			createResp, createErr := servicesClient.Create(&services.CreateServiceInput{
+				FriendlyName: "Test Service",
+			})
+			Expect(createErr).To(BeNil())
+			Expect(createResp).ToNot(BeNil())
+			Expect(createResp.Sid).ToNot(BeNil())
+
+			pageResp, pageErr := servicesClient.Page(&services.ServicesPageOptions{})
+			Expect(pageErr).To(BeNil())
+			Expect(pageResp).ToNot(BeNil())
+			Expect(len(pageResp.Services)).Should(BeNumerically(">=", 1))
+
+			paginator := servicesClient.NewServicesPaginator()
+			for paginator.Next() {
+			}
+
+			Expect(paginator.Error()).To(BeNil())
+			Expect(len(paginator.Services)).Should(BeNumerically(">=", 1))
+
+			serviceClient := verifySession.Service(createResp.Sid)
+
+			fetchResp, fetchErr := serviceClient.Fetch()
+			Expect(fetchErr).To(BeNil())
+			Expect(fetchResp).ToNot(BeNil())
+
+			updateResp, updateErr := serviceClient.Update(&service.UpdateServiceInput{})
+			Expect(updateErr).To(BeNil())
+			Expect(updateResp).ToNot(BeNil())
+
+			deleteErr := serviceClient.Delete()
+			Expect(deleteErr).To(BeNil())
+		})
+	})
+})