+++ service/verify/tests/unit/verify_suite_test.go
@@ -1,0 +1,15 @@
+// +build unit
+
+package tests
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestVerifyV2(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Verify V2 Test Suite")
+}